examples/integrations/portia/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.52.1
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
boto3==1.38.25
botocore==1.38.25
browser-use==0.2.5
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Cython==3.1.1
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.16
et_xmlfile==2.0.0
faiss-cpu==1.11.0
filetype==1.2.0
frozenlist==1.6.0
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.40.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.73.0rc1
grpcio-status==1.73.0rc1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
instructor==1.8.3
Jinja2==3.1.6
jiter==0.8.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
langchain==0.3.22
langchain-anthropic==0.3.3
langchain-aws==0.2.19
langchain-core==0.3.49
langchain-deepseek==0.1.3
langchain-google-genai==2.1.2
langchain-ollama==0.3.0
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langgraph==0.2.76
langgraph-checkpoint==2.0.26
langgraph-sdk==0.1.70
langsmith==0.3.43
loguru==0.7.3
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
mcp==1.9.2
mdurl==0.1.2
mem0ai==0.1.93
monotonic==1.6
multidict==6.4.4
numpy==1.26.4
ollama==0.4.9
openai==1.82.1
openpyxl==3.1.5
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.2.3
patchright==1.52.4
playwright==1.52.0
pluggy==1.6.0
portalocker==2.10.1
portia-sdk-python==0.3.9
posthog==3.25.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic-settings==2.9.1
pydantic_core==2.27.2
pyee==13.0.0
Pygments==2.19.1
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyperclip==1.9.0
pytest==8.3.5
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.14.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9.1
s3transfer==0.13.0
screeninfo==0.8.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.3.5
starlette==0.47.0
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uuid7==0.1.0
uvicorn==0.34.2
yarl==1.20.0
zstandard==0.23.0
